refactor(vendo): wire route table — automations/runs (kill-list B4)
